+++ src/test/auth.spec.ts
@@ -18,13 +18,13 @@ describe('Auth', () => {
 
   describe('auth errors', () => {
 
-    // ensure token is fetched
-    removeTokenFile();
-
     it('should forward error', done => {
+      // ensure token is fetched
+      removeTokenFile();
+
       const client = new CoreAPIClient({ ...integrationTestConfig, debug: false, clientIdentifier: 'i/n/v/a/l/i/d', clientSecret: '******' });
       client.query(`does not matter`, ['ServiceCall'])
-        .catch((errorResp: ErrorResponse<{ error: string, error_description: string }>) => {
+        .catch((errorResp: ErrorResponse<{ error: string, error_description: string }, any>) => {
           assert.strictEqual(errorResp.statusCode, 401);
           assert.strictEqual(errorResp.message, 'Unauthorized');
           assert.strictEqual(errorResp.error.error, 'invalid_client');
@@ -32,6 +32,39 @@ describe('Auth', () => {
           return done();
         });
     }).timeout(5000);
+
+    it('should forward error for password', done => {
+
+      // ensure token is fetched
+      removeTokenFile();
+
+      const client = new CoreAPIClient({ ...integrationTestConfig, debug: false, authGrantType: 'password', authUserName: 'i/n/v/a/l/i/d', authPassword: '******' });
+      client.query(`does not matter`, ['ServiceCall'])
+        .catch((errorResp: ErrorResponse<{ error: string, error_description: string, disassociated: string }, any>) => {
+          assert.strictEqual(errorResp.statusCode, 400);
+          assert.strictEqual(errorResp.error.error, 'invalid_grant');
+          assert.strictEqual(errorResp.error.error_description, 'Bad credentials');
+          assert.strictEqual(errorResp.error.disassociated, 'true');
+          return done();
+        });
+    }).timeout(5000);
+
+
+    it('should forward error unknown endpoints', done => {
+
+      // ensure token is fetched
+      removeTokenFile();
+
+      const client = new CoreAPIClient({ ...integrationTestConfig, debug: false, oauthEndpoint: 'https://ds.coresuite.com/api/RANDOM-API' });
+      client.query(`does not matter`, ['ServiceCall'])
+        .catch((errorResp: ErrorResponse<{ error: string, error_description: string, disassociated: string }, any>) => {
+          assert.strictEqual(errorResp.statusCode, 404);
+          assert(!!errorResp.error);
+          return done();
+        });
+    }).timeout(5000);
+
+
   });