+++ src/test/batch-api.spec.ts
@@ -1,7 +1,7 @@
 import assert = require('assert');
 import { integrationTestConfig } from './integration-test.config';
 import { CoreAPIClient } from '../core-api.client';
-import { CreateAction, DeleteAction } from '../core/batch-action.model';
+import { CreateAction, DeleteAction } from '../index';
 
 describe('BatchApi', () => {