+++ src/test/integration-test.config.ts
@@ -1,5 +1,6 @@
+import { readFileSync } from 'fs';
 import { ClientConfig } from '../core/client-config.model';
-
+const packageJson = JSON.parse(readFileSync('./package.json').toString()) as { version: string };
 require('dotenv').config()
 
 export const integrationTestConfig: ClientConfig = {
@@ -11,7 +12,7 @@ export const integrationTestConfig: ClientConfig = {
 
   clientIdentifier: process.env.CLIENT_IDENTIFIER as string,
   clientSecret: process.env.CLIENT_SECRET as string,
-  clientVersion: process.env.CLIENT_VERSION as string,
+  clientVersion: packageJson.version as string,
 
   authGrantType: process.env.AUTH_GRANT_TYPE as 'password' | 'client_credentials' || 'password',
 
@@ -21,3 +22,15 @@ export const integrationTestConfig: ClientConfig = {
   authCompany: undefined // use first 
 
 };
+
+Object.keys(integrationTestConfig).forEach((key) => {
+  const missingMandatoryValue = [
+    'clientIdentifier',
+    'clientSecret',
+    'clientVersion'
+  ].indexOf(key as keyof ClientConfig) !== -1 && !integrationTestConfig[key as keyof ClientConfig];
+
+  if (missingMandatoryValue) {
+    throw new Error(`missing mandatory config key [${key}]\n`);
+  }
+}, true)