+++ CHANGELOG.md
@@ -5,9 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] 2021-07-26
+### Changed
+- **BREAKING** Dropping support for Node.js < v12, sdk now requires Node.js v12 or newer 
+### Added
+- Adding support for using client in browser (isomorphic), migrate from `node-fetch` to `isomorphic-fetch`
+- **BREAKING** Adding `user_id` to `OauthTokenResponse`
+- Adding compatibility tests
+### Updated 
+- Updating down stream npm packages 
+
+
 ## [1.2.4] 2021-03-24
 ### Updated 
-- fix error handling, http errors 4xx/5xx status throw ErrorResponse, 304 and 302 do not throw 
+- Fix error handling, http errors `4xx/5xx` status throw `ErrorResponse`, `304` and `302` do not throw 
 
 ## [1.2.3] 2021-03-16
 ### Changed 
@@ -15,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.2.2] 2021-01-15
 ### Added
-- batch api support 
+- Batch api support 
 ### Updated 
 - DTO versions, removed CrowdBusinessPartner, CrowdAssignment, CrowdPerson DTO
 ### Changed